requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
pyiceberg==0.5.0
boto3==1.34.0
pytest==8.0.0
black==24.1.0
flake8==7.0.0
isort==5.13.2
pandas==2.2.0
apache-airflow==2.8.1
apache-airflow-providers-sqlite==3.4.0
python-dotenv==1.0.1
minio
